Use role from token and combine templates
Version 28.0.0 of the utils repo means that if a token is expired it
returns the role type that's being created rather than nothing.

This allows us to drop the role_param from the url and move to just
having one url for creating buyers and suppliers.

The templates have also been combined to reduce replication.

If the token supplied is expired, we render a role specific error page.
If the token is None (invalid), we render a 400 error page with a custom
error message