Use one form for creating buyers and suppliers
The CreateUserForm can be used for buyers and suppliers. The only
different field is phone number for buyers. This is an optional field,
and so we can use the same form for suppliers. There is some logic in
the create buyer template which prevents the phone number field being
rendered for suppliers.