Add a check for safe url for redirects after login
After the buyer users have been added to the list of roles the prefix
checks no longer restrict protocol-relative URLs (eg //example.com),
allowing the login page to redirect the user to an external domain.

This is adding a `is_safe_url` function based on the one from
http://flask.pocoo.org/snippets/62/ (it's also used by Notify).

The function is modified to return `False` for empty target URLs.

This also removes per-role redirect restrictions since their original
purpose was to prevent the open redirects. User role now only affects
the default redirect location if next_url was not set, otherwise
next_url is used as is.