Replace Mandrill config variables with Notify
Adds a config value for Notify API key and a list of notify templates.

We're splitting templates between preview and staging/production so
that we can test new versions of templates before rolling them out
and this enables us to set up production templates to not store any
personalisation data after the email has been sent.

NOTIFY_TEMPLATES is a dictionary to group templates together, so to
make sure we're not missing any of the templates in staging and prod
there's an assert in Staging config to check that it has the same
keys as the default/preview config.