Remove `_index-page.scss`
Although we are using the `index-page` class on some of our
templates, there would only be styles in use if we used some of
the nested styles that apply to html elements inside an
`index-page` class e.g. an `h2` inside a div with class
`index-page`. Because of that fact, we can remove all the nested
styles for the `index-page` class and once those are all removed,
the `index-page` class itself does nothing so is removed from both
our scss and our templates.