Add back in `break-links` class but move it to _text.scss
We removed `.break-links` from the `application.scss` file as we
thought it was not needed. However it turns out that html is
generated using dm-utils that uses this class whenever we call the
`format_links` jinja filter. Whilst we are not actually calling
`format_links` in this application, it feels like something that
may be done at any time and it wouldn't feel responsible to leave
it to the developer adding the filter to know to also write the
scss for it.

The `text_scss` file also seems more appropriate for this to live.