Use `SHARED_EMAIL_KEY` to encrypt reset password email tokens
We will be changing our tokens for reset password emails to be
generated using the `SHARED_EMAIL_KEY` rather than the `SECRET_KEY`.
This is because
- this will be consistent with the other email tokens we generate
- although there is technically no harm in using `SECRET_KEY`,
the actual purpose of the `SECRET_KEY` is to be used for encrypting
Flask sessions. This would mean if we needed to rotate a key for
our Flask sessions it would also break out email tokens and vice
versa. By using different keys we should reduce the risk of a
token rotation having an impact on too large a part of the site

Note, the new version of utils brought in will mean that we still
succesfully decrypt and allow tokens that were previously
generated using the old `SECRET_KEY`. After 1 full day of support,
they would all become invalid anyway so after that point we can
also remove support for allowing old `SECRET_KEY` tokens.

By bringing in the new version of utils I also ran
`make freeze-requirements` and allowed a few non breaking changes
of other dependancies. Strangely though, a non breaking change
increase of `Werkzeug` to 0.14 actually breaks our unit tests so
I manually removed that bump from `requirements.txt`.