Bump Werkzeug and fix test relying on now removed Werkzeug bug
`freeze-requirements` from our previous commit also bumps
Werkzeug to 0.14 however doing so broke one of our unit tests. It
turns out this is because we were asserting values returned from
`Werkkzeug.parse_cookie` that should not have been returned.
`parse_cookie` is only supposed to return a dict of values where
the keys do not match the core cookie directives
(https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie)
but instead we were getting `Secure; HttpOnly; Path` as a key.
This is likely linked to a bug with values being a single `/`
(for our case, the value of `Path`). This behaviour was fixed in
0.14 and now these core directives don't get returned and instead
we need to look at the header to check they are set.