notify example redirection: switch the target email addresses to SES's equivalent simulated addresses
this should work fine as long as notify is using SES as their backend. if they
change this, we just have to change this setting. using this has the advantage
that it will appear to be "delivered" as far as notify's concerned and a
record of the message's reference id will be kept.