Log and redirect with success flash message if alert email fails
The user should be told that their password change was successful,
even if the alert email failed to send. Instead of raising a 5xx
error, just log an error message.

The logging itself is done in a new helper so we can assert that
it's been called without affecting the test client.