Add validation masthead to password reset page
The password reset page previously did not use the same type of form as
the login page. This meant that the form would rely on browser
validation messages which are not as visually identifiable as the ones
from the dm toolkit.

We now get a validation error masthead with message when the form is
submitted with an email or with an invalid email, the same as our login
page.