Bump digitalmarketplace-utils dependency to v48.2.0
also bump digitalmarketplace-content-loader to 5.2.1 and other version
bumps that come with the re-freeze. this includes werkzeug 0.15, which
requires a couple of minor changes to our tests that depend on the default
behaviour of werkzeug/flask, which has changed but not in a way that I
think is particularly important to us.