Replace custom `pip freeze` script with `pip-tools`
In the past we've avoided using out-of-the-box solutions for Python
dependency resolution because a) they haven't been very mature and b)
we've had lots of issues with version conflicts. See [[1]], [[2]] for
details. Instead, we've been using a custom Python script that
under-the-hood runs `pip freeze` and saves the output to
`requirements.txt`.

This script works well for us, but it doesn't integrate well with other
tools. On the other hand [`pip-tools`](https://github.com/jazzband/pip-tools)
as of 2020 seems to be well-supported by its maintainers and other
tools; for instance, GitHub's automated update service
[Dependabot](https://dependabot.com) supports `requirements.in` files.

At the same time, we've been better at keeping our dependencies
up-to-date and fixing version conflicts, so there is no longer any issue
caused by using `pip-tools` to generate our `requirements.txt` file. I
think it's time to give `pip-tools` another chance :)

This commit replaces our script `dmutils.repoutils.freeze_requirements`
with `pip-compile`, and changes our `requirements-app.txt` file to be
`requirements.in` to meet `pip-tools` convention. There are also
separate `requirements-dev.in` and `requirements-dev.txt` files for dev
requirements, following the [layered requirements example in the
`pip-tools` README][3].

[1]: https://gds-way.cloudapps.digital/manuals/programming-languages/python/python.html#applications
[2]: https://github.com/alphagov/digitalmarketplace-api/commit/95ac12206d26e6b219dd381dd63641c33467afbd
[3]: https://github.com/jazzband/pip-tools#workflow-for-layered-requirements